server/requirements.txt
# server/requirements.txt

grpcio>=1.62.3
grpcio-tools>=1.62.3
grpcio-health-checking>=1.62.3

opentelemetry-sdk>=1.25.0
opentelemetry-exporter-prometheus>=0.46b0

# Prometheus 웹 서버를 시작하기 위한 라이브러리 명시
prometheus-client>=0.20.0

opentelemetry-instrumentation-grpc
